Remove watcher from context
